refactor: add validation for ID in delete method
- Added validation for the ID in the delete method to ensure it is a valid UUID before deleting the record.

+++ api/src/app/controllers/ContactController.js
@@ -101,6 +101,10 @@ class ContactController {
         // deletar registro
         const { id } = request.params;
 
+        if (!isValidUUID(id)) {
+            return response.status(400).json({ error: "Invalid contact" });
+        }
+
         await ContactRepository.delete(id);
         response.sendStatus(204);
     }
